==> 7_NYT_corpus.ipynb <==
import os
import re
from urllib import urlopen
from nltk import word_tokenize
from nltk import bigrams
from nltk import trigrams
from nltk import ngrams

from nltk.stem import PorterStemmer
pt = PorterStemmer()

from nltk.tag.stanford import NERTagger

st = NERTagger('/Users/rterman/Dropbox/berkeley/Programming-resources/stanford-ner-2014-10-26/classifiers/english.all.3class.distsim.crf.ser.gz',
                '/Users/rterman/Dropbox/berkeley/Programming-resources/stanford-ner-2014-10-26/stanford-ner.jar') 

### Prepare stop words

stop_words = urlopen('http://jmlr.org/papers/volume5/lewis04a/a11-smart-stop-list/english.stop').read().split('\n')

# apply porter stemmer to stop words
stop_words = map(pt.stem,stop_words)
stop_words.append('x')

dir = "/Users/rterman/Dropbox/berkeley/Dissertation/Data and Analyais/Git Repos/human-rights-coverage/Data/New York Times/Texts"

#importing docs and pre-processing them all in one!

docs = {}
for file_name in os.listdir(dir):
    if file_name.endswith(".txt"):
        dict = {}
        name = file_name[:-4]
        ## tag region
        if 'africa' in file_name:
            region = 'africa'
            number = file_name[6:-4]
        if 'mena' in file_name:
            region = 'mena'
            number = file_name[4:-4]
        if 'la' in file_name:
            region = 'la'
            number = file_name[2:-4]
        if 'eeca' in file_name:
            region = 'eeca'
            number = file_name[4:-4]
        if 'asia' in file_name:
            region = 'asia'
            number = file_name[4:-4]
        if 'west' in file_name:
            region = 'west'
            number = file_name[4:-4]
        dict['region'] = region
        dict['number'] = number
        f = open(dir + '/' + file_name)
        i = f.read() # read in text

        """
        ## get rid of proper nouns
        i = st.tag(i.split())
        for term in i:
            if term[1]=='PERSON' or term[1]=='LOCATION' or term[1]=='ORGANIZATION':
                i.remove(term)
        i = [x[0] for x in i]
        i = ' '.join(i)
        """
        ## preprocess
        i = re.sub('\W', ' ', i) # remove punctuation
        i = i.lower() # remove capitalization
        i = word_tokenize(i) # tokenize the words
        i = map(pt.stem, i) # apply porter stemmer 
        i = [x for x in i if x not in stop_words] # remove stop words
        dict['text'] = i
        f.close()
        docs[name] = dict

len(docs)

docs['africa100']['text']

# count number of unigrams
unigrams = {}
used = []

for doc in docs:
    for word in docs[doc]['text']:
        if word in unigrams:
                unigrams[word] += 1 
        if word not in unigrams and word not in used:
                unigrams[word] = 1 # add the word to unigrams for the first time
                used.append(word) # add the word to used, which is just a list of the words used
len(set(unigrams)) # the number of the unique unigrams

from heapq import nlargest
import operator

top_uni_nu = nlargest(5000, unigrams.iteritems(), key=operator.itemgetter(1))
top_uni = [x[0] for x in top_uni_nu]

top_uni_nu[-10:-1] # what are the bottom 10 unigrams?

top_uni[1]

** playing around with nertagger**

tagged = st.tag(top_uni)

tagged[1:10]

top_uni_nopn = []
for i in tagged:
    if i[1] == 'O':
        top_uni_nopn.append(i[0])
    

len(top_uni_nopn)

# create lists of counts for unigrams
uni_dtm = []
for doc in docs:
    vec = [doc]
    # make a list of unigrams in that statement and count them
    for i in top_uni_nopn:
        count = docs[doc]['text'].count(i)
        vec.append(count)
    uni_dtm.append(vec)   

**Create a N × 500 document term matrix X, where the columns count the unigrams
and the rows are the documents**

dtm = open('/Users/rterman/Dropbox/berkeley/Dissertation/Data and Analyais/Git Repos/human-rights-coverage/Data/New York Times/uni_dtm.csv', 'w')
row = 'doc'
for i in top_uni_nopn:
    row = row + ',' + i
dtm.write(row + '\n')

# write rows

for i in range(len(uni_dtm)):
    x = str(uni_dtm[i])
    x = x.replace('[','')
    x = x.replace(']','')
    dtm.write(x + '\n')

dtm.close()
